Revert "Merge Feature/code enhancement (#17)"
This reverts commit de3294aa840bae596427c70e9ed5e7ca36c73fa1.
