+++ jsconfig.json
@@ -1,30 +1,15 @@
 {
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "@zede-components/*": ["src/components/*"],
-      "@zede-hooks/*": ["src/hooks/*"],
-      "@zede-utils/*": ["src/utils/*"],
-      "@zede-static/*": ["src/static/*"],
-      "@zede-services/*": ["src/services/*"],
-      "@zede-public/*": ["public/*"]
-    },
-    "jsx": "preserve",
-    "outDir": "./build",
-    "rootDir": "./",
-    "strict": true,
-    "esModuleInterop": true,
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "allowSyntheticDefaultImports": true,
-    "noFallthroughCasesInSwitch": true,
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "noEmit": true
-  },
-  "include": ["src"],
-  "exclude": ["node_modules"]
+    "compilerOptions": {
+        "baseUrl": ".",
+        "paths": {
+            "@zede/*": ["*"],
+            
+            "@zede-components/*": ["src/components/*"],
+            "@zede-hooks/*": ["src/hooks/*"],
+            "@zede-utils/*": ["src/utils/*"],
+            "@zede-static/*": ["src/static/*"],
+            "@zede-services/*": ["src/services/*"],
+            "@zede-public/*": ["public/*"],
+        }
+    }
 }
