+++ package.json
@@ -29,7 +29,6 @@
     "next-fonts": "^1.5.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-hook-form": "^7.17.0",
     "react-lottie-player": "^1.4.0",
     "react-material-ui-carousel": "^2.1.1",
     "react-svg": "^14.0.13"
