+++ src/components/LandingPage/Hero.js
@@ -73,11 +73,6 @@ const useStyle = makeStyles((theme) => ({
     height: "100vh",
     minHeight: "700px",
     position: "absolute",
-    overflow: "hidden",
-  },
-  anim: {
-    width: "100%",
-    height: "100%",
     opacity: 0.5,
     zIndex: 0,
     top: 0,
@@ -86,15 +81,20 @@ const useStyle = makeStyles((theme) => ({
     strokeWidth: 0,
     strokeOpacity: 0.5,
     fill: theme.palette.secondary.dark,
-    [theme.breakpoints.down("xs")]: {
-      transform: "scale(2)",
-    },
   },
 }));
 
 const Hero = (params) => {
   const classes = useStyle();
   const router = useRouter();
+  // const animOptions = {
+  //   loop: true,
+  //   autoplay: true,
+  //   animationData: animationData.default,
+  //   rendererSettings: {
+  //     preserveAspectRatio: "xMidYMid slice",
+  //   },
+  // };
 
   return (
     <Paper className={classes.wrapper}>
@@ -144,8 +144,11 @@ const Hero = (params) => {
         <Lottie
           loop
           play
-          className={classes.anim}
           animationData={animationData}
+          style={{
+            width: "100%",
+            height: "900px",
+          }}
         />
       </div>
     </Paper>
