+++ src/utils/globalFonts.js
